% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{close_connections}
\alias{close_connections}
\title{Close all connections and remove them from the connections list}
\usage{
close_connections()
}
\description{
Close all connections and remove them from the connections list
}
