% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{combine_configs}
\alias{combine_configs}
\title{Combine the optionally nested lists root and new.}
\usage{
combine_configs(root, new)
}
\arguments{
\item{root}{The existing yml config}

\item{new}{The yml to be added}
}
\value{
Combined yml config

Elements that exists in new but not root will be added to root
Elements that exists in root but not new will be retained in root
Elements that exist in both will be replaced in root by the contents of new
}
\description{
Combine the optionally nested lists root and new.
}
