% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_examples.R
\name{config_examples}
\alias{config_examples}
\title{Examples of yaml configurations for database connections}
\usage{
config_examples(filename = NA)
}
\arguments{
\item{filename}{A string. If supplied, examples are written to a file with this name.}
}
\value{
A yaml string of database configuration examples, invisibly.
}
\description{
Provides example configurations for several databases and a range of options
}
\details{
Irrespective of whether a filename is supplied, yaml configuration examples
will be returned invisibly as a single string and printed if the session is
interactive.
}
\examples{
config_examples()

if(FALSE){
    # to write the examples to a file called 'examples.yml'
    config_examples("examples.yml")
}

}
