% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{(Re-)Establish connections to databases}
\usage{
connect(.config_filename = NA, exclusive = FALSE)
}
\arguments{
\item{exclusive}{if .fn is present, should it be used exclusively (TRUE) or
combined with user and site configs (FALSE)?
Closes and then re-opens all configured connections}

\item{.fn}{String, the name of a config file}
}
\description{
(Re-)Establish connections to databases
}
