% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_info.R
\name{connection_info}
\alias{connection_info}
\title{Get information about available database connection}
\usage{
connection_info(search = ".*")
}
\arguments{
\item{search}{A regular expression to be used identify connection names to
include. The default ('.*') returns all of them.}
}
\value{
A tibble with the following columns:

\if{html}{\out{<div class="sourceCode">}}\preformatted{* name - identifier (character)
* description - a description of the connection, if found in the conf file (character)
* driver - the name of the driver used (character)
* conn_str - the string used to parameterize the connection (character)
* pool - is this a pool connection? (boolean)
}\if{html}{\out{</div>}}
}
\description{
Get information about available database connection
}
