% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connections_list}
\alias{connections_list}
\title{Return a list of available connection names}
\usage{
connections_list()
}
\value{
A list of connections maintained by the loaded sqlhelper that are
currently live.
}
\description{
Return a list of available connection names
}
