% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{create_connections}
\alias{create_connections}
\title{Populate the list of available connections}
\usage{
create_connections(config_name = NA, exclusive = FALSE)
}
\arguments{
\item{config_name}{String. Name of a config file. Defaults to NA. See
\code{\link[=read_configs]{read_configs()}} for details.}

\item{exclusive}{Boolean. Should this file be used exclusively to define
connections? Defaults to \code{FALSE}. See \code{\link{read_all_configs}}.}
}
\description{
Populate the list of available connections
}
