% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_utils.R
\name{driver}
\alias{driver}
\title{Determine the connection driver to use}
\usage{
driver(conf)
}
\arguments{
\item{conf}{A named list representing a single connection returned by
\code{\link{validate_configs}}.}
}
\value{
A driver function. Defaults to \code{\link[odbc]{odbc}} if no
"server_type" element is found.
}
\description{
Determine the connection driver to use
}
\details{
Search for an element named "server_type" and return a driver
function appropriate for that database. For example, "server_type:
sqlite" in your config will cause \code{driver()} to return
\code{\link[RSQLite]{SQLite}}.
}
