% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sql.R
\name{extract_create_statements_efficient}
\alias{extract_create_statements_efficient}
\title{Extract CREATE statements using line-by-line parsing}
\usage{
extract_create_statements_efficient(
  file_path,
  statement_types = c("TABLE", "VIEW", "INDEX", "TRIGGER", "FUNCTION", "PROCEDURE")
)
}
\arguments{
\item{file_path}{Path to the SQL dump file}

\item{statement_types}{Vector of CREATE statement types to extract}
}
\value{
Tibble with CREATE statements
}
\description{
Extract CREATE statements using line-by-line parsing
}
