% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{extract_quoted_strings}
\alias{extract_quoted_strings}
\title{Extract and cache single quoted strings from lines of sql}
\usage{
extract_quoted_strings(lines)
}
\arguments{
\item{lines}{A character vector containing consecutive lines of SQL code.}
}
\description{
Extract and cache single quoted strings from lines of sql
}
