% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{get_all_configs}
\alias{get_all_configs}
\title{Return the combined available configurations}
\usage{
get_all_configs(.fn = NA, exclusive = FALSE)
}
\arguments{
\item{.fn}{String. The full name and path of a configuration file.}

\item{exclusive}{Boolean. If TRUE, the file named by the .fn parameter is
treated as the only config file. Site and user level files are not read.
This parameter is ignored if .fn is missing.}
}
\value{
optionally nested named list or vector as returned by
\code{yaml::read_yaml}
}
\description{
Return the combined available configurations
}
