% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{get_cache_env}
\alias{get_cache_env}
\title{Access the environment containing the connections cache}
\usage{
get_cache_env()
}
\value{
a sibling env to sqlhelper
}
\description{
Access the environment containing the connections cache
}
