% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{get_config}
\alias{get_config}
\title{Return the contents of a config file}
\usage{
get_config(f = NA)
}
\arguments{
\item{String, }{either the full path and name of a config file, or one of
'site' or 'user'. If omitted, \code{r get_config()} will search the current
directory for a config file.}
}
\value{
optionally nested named list or vector as returned by
\code{r yaml::read_yaml}
}
\description{
Return the contents of a config file
}
