% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlhellper.r
\name{get_sql}
\alias{get_sql}
\title{Returns the SQL}
\usage{
get_sql(
  keys,
  params = NULL,
  .multiline = FALSE,
  config_path = system.file("ini/sql.toml", package = "sqlhelper")
)
}
\arguments{
\item{keys}{a slash separated key string (eg: "section/query")}

\item{params}{a named list with the params to substitue in the query}

\item{.multiline}{is the query a list of statements? then place a \code{TRUE} here}

\item{config_path}{path to the INI/TOML file where SQL is stored}
}
\value{
a string containing the SQL query
}
\description{
Returns the SQL
}
\examples{
  get_sql("simple/SELECT_TEST0")
  get_sql("whisker/query", list(name = "Dwayne"))
}
