% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_utils.R
\name{getrunparams}
\alias{getrunparams}
\title{Returns a connection and a sql runnner for the db parameter. For internal use
only!}
\usage{
getrunparams(conn_name)
}
\arguments{
\item{conn_name}{the name of a connection}
}
\value{
A list, with elements 'conn' (the connection), 'runner' (the run
function), and 'is.live' (a validity test for conn), or NULL if conn_name
is not in the connections cache
}
\description{
For now, sqlhelper only uses DBI so this isn't really necessary, but if in
the future we need functions from other packages we can add them here without
needing to touch anything else.
}
