% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sql.R
\name{interpolate_sql}
\alias{interpolate_sql}
\title{Parameterize SQL queries}
\usage{
interpolate_sql(
  interpolate,
  quotesql,
  conn_name,
  sql,
  values = parent.frame(),
  default_conn = NULL
)
}
\arguments{
\item{sql}{A tibble, as returned by \code{\link[=read_sql_file]{read_sql_file()}}}

\item{default_conn}{A connection inheriting from DBI::DBIConnection. Used
when the \code{conn} variable of the \code{sql} argument is set to
'default'.}

\item{env}{An environment containing values for interpolation. Defaults to
the value of \code{\link[=parent.frame]{parent.frame()}}}
}
\value{
A list or character vector of parameterized sql queries
}
\description{
Accepts a list or character vector of SQL queries and parameterizes each.
}
