% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{is.connected}
\alias{is.connected}
\title{Test whether a database is connected}
\usage{
is.connected(db)
}
\arguments{
\item{db}{Which database do you want to test? (restricted to 'cds' in MHCLG)}
}
\value{
Boolean

is.connected and not.connected provide generic, semantically transparent
methods of checking whether a connection is live, irrespective of whether it
is through RODBC, RPostgreSQL, DBI etc. They depend on the is.live functions
provided by getrunparams (see above).
is.connected returns FALSE if the db connection returned by getrunparams is
not valid.
}
\description{
Test whether a database is connected
}
