% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{new_connection_cache}
\alias{new_connection_cache}
\title{Create the connection cache}
\usage{
new_connection_cache()
}
\value{
The new connection cache list, invisibly
}
\description{
Create the connection cache
}
