% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlgenerators.R
\name{nthofk}
\alias{nthofk}
\title{Restrict a query to a sample of 1/nth of the unique values of a key}
\usage{
nthofk(n, k)
}
\arguments{
\item{n}{nth of k}

\item{k}{key on which to restrict sample}
}
\value{
string (sql code)
}
\description{
Restrict a query to a sample of 1/nth of the unique values of a key
}
\seealso{
Other SQL Generators: 
\code{\link{vinnthoffk}()}
}
\concept{SQL Generators}
