% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sql.R
\name{parse_sql_to_csv}
\alias{parse_sql_to_csv}
\title{Parse SQL dump line by line and stream to CSV file}
\usage{
parse_sql_to_csv(
  sql_file_path,
  csv_file_path,
  sql_keywords = c("CREATE", "INSERT", "UPDATE", "DELETE", "ALTER", "DROP"),
  start_line = 1,
  append = FALSE,
  batch_size = 1000
)
}
\arguments{
\item{sql_file_path}{Path to the SQL dump file}

\item{csv_file_path}{Path to output CSV file}

\item{sql_keywords}{Vector of SQL keywords to look for (default: common DML/DDL)}

\item{start_line}{Line number to start processing from (for resuming)}

\item{append}{Whether to append to existing CSV file (TRUE for resuming)}

\item{batch_size}{Number of statements to collect before writing to CSV}
}
\value{
Number of statements processed
}
\description{
Parse SQL dump line by line and stream to CSV file
}
