% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sql.R
\name{prepare_sql}
\alias{prepare_sql}
\title{prepare to queries and meta data for running}
\usage{
prepare_sql(
  sql,
  quotesql = "yes",
  values = parent.frame(),
  execmethod = "get",
  geometry = NA,
  default_conn = NA
)
}
\arguments{
\item{sql}{An optionally-named list or character vector containing sql
commands, or a tibble returned by \code{\link[=read_sql]{read_sql()}}}

\item{quotesql}{"yes" or "no" - should interpolated characters be quoted by
default?}

\item{values}{Should the SQL be parameterized from R? Defaults to the
value of \code{parent.frame()}. Pass any object that is not an environment
(e.g. "no" or FALSE) if interpolation is to be skipped, or another
environment containing values to interpolate to avoid using
\code{.GlobalEnv}.}

\item{execmethod}{One of "get", "send" or "spatial" - which method should be
used to execute the query? "get" means \code{\link[DBI:dbGetQuery]{DBI::dbGetQuery()}}; "send" means
\code{\link[DBI:dbSendQuery]{DBI::dbSendQuery()}}; "spatial" means \code{\link[sf:st_read]{sf::st_read()}}.}

\item{geometry}{If \code{execmethod} is "spatial", which column contains the
geometry? (ignored if \code{execmethod} is not spatial)}

\item{default_conn}{Either the name of a sqlhelper connection, or a database
connection returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}, or NA}
}
\description{
prepare to queries and meta data for running
}
