% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect.R
\name{prune}
\alias{prune}
\title{remove a connection from the connections cache}
\usage{
prune(conn_name)
}
\arguments{
\item{conn_name}{The name of the connection to be removed}
}
\description{
remove a connection from the connections cache
}
