% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{read_config_file}
\alias{read_config_file}
\title{Return the contents of a config file}
\usage{
read_config_file(conf = NA)
}
\arguments{
\item{conf}{String. Either the path and name of a config file, or one
of 'site' or 'user', or \code{NA} (the default). If \code{NA},
\code{get_config} will search the current directory for a config file.}
}
\value{
optionally nested named list or vector as returned by
  \code{yaml::read_yaml}, or \code{NA} if the appropriate file does not
  exist.
}
\description{
Return the contents of a config file
}
\details{
The default search name for config files is "sqlhelper_db_conf.yml".
  If \code{conf} is 'site', 'user' or \code{NA}, a file with this name will
  be sought in \code{rappdirs::site_config_dir},
  \code{rappdirs::user_config_dir} or the current working directory,
  respectively. In an interactive context, a warning will be issued if the
  sought file does not exist.
}
