% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_configs.R
\name{read_configs}
\alias{read_configs}
\title{Return the combined available configurations}
\usage{
read_configs(config_name = NA, exclusive = FALSE)
}
\arguments{
\item{exclusive}{Boolean. If TRUE, the file named by the .config_filename parameter is
treated as the only config file. Site and user level files are not read.
This parameter is ignored if .fn is missing.}

\item{.config_filename}{String. The full name and path of a configuration file.}
}
\value{
optionally nested named list or vector as returned by
\code{yaml::read_yaml}
}
\description{
Return the combined available configurations
}
