% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{reinstate_quoted_strings}
\alias{reinstate_quoted_strings}
\title{Replace tokens with their matched quoted strings in lines of sql}
\usage{
reinstate_quoted_strings(lines, id)
}
\arguments{
\item{lines}{A character vector containing consecutive lines of SQL code.}
}
\description{
Replace tokens with their matched quoted strings in lines of sql
}
