% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sql.R
\name{resume_sql_parsing}
\alias{resume_sql_parsing}
\title{Resume SQL parsing from a specific line}
\usage{
resume_sql_parsing(
  sql_file_path,
  csv_file_path,
  resume_line,
  sql_keywords = c("CREATE", "INSERT", "UPDATE", "DELETE", "ALTER", "DROP"),
  batch_size = 1000
)
}
\arguments{
\item{sql_file_path}{Path to the SQL dump file}

\item{csv_file_path}{Path to output CSV file}

\item{resume_line}{Line number to resume from}

\item{sql_keywords}{Vector of SQL keywords to look for}

\item{batch_size}{Number of statements to collect before writing to CSV}
}
\value{
Number of statements processed
}
\description{
Resume SQL parsing from a specific line
}
