% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlrunners.R
\name{runfile}
\alias{runfile}
\title{execute a \strong{single} file of SQL queries and obtain a flat list of results.}
\usage{
runfile(fn, conn = NA, interpolate = parent.frame())
}
\arguments{
\item{fn}{is a file name.}

\item{conn}{is passed from runfiles and defaults to NA - the preferred method, for transparency,
is to use an interpreted comment inside the sql file.}

\item{interpolate}{defaults to the value of \code{parent.frame()}. May be set to \code{FALSE} if
interpolation to be avoided, or to another environment to control the source of the used. This
can be useful if sqlhelper is used to run SQL from within another package, and you don't want
to interfere with \code{.GlobalEnv}.}
}
\value{
If the file contains more than one query, a list. Each element contains the results of
each query.

\if{html}{\out{<div class="sourceCode">}}\preformatted{   If the file contains a single query, the result of that query is returned.
}\if{html}{\out{</div>}}
}
\description{
This function is used internally and not exported, use \code{\link{runfiles}} instead.
}
