% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlrunners.R
\name{runfiles}
\alias{runfiles}
\title{Read, prepare and execute .SQL files}
\usage{
runfiles(filenames, ...)
}
\arguments{
\item{filenames}{name, or vector of names, of file(s) to be executed}

\item{...}{default parameters for \code{\link[=runqueries]{runqueries()}}}
}
\value{
A list of results of sql queries found in files
}
\description{
Accepts a character vector of SQL file names and attempts to execute each one.
}
