% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{set_connections}
\alias{set_connections}
\title{Populate the list of available connections}
\usage{
set_connections(config_filename = NA, exclusive = FALSE)
}
\arguments{
\item{config_filename}{String. Name of a config file. Defaults to NA. See
\code{\link{read_config_file}} and \code{\link{get_all_configs}} for
options.}

\item{exclusive}{Boolean. Should this file be used exclusively to define
connections? Defaults to \code{FALSE}. See \code{\link{get_all_configs}}.}
}
\description{
Populate the list of available connections
}
\details{
This function is called by \code{.onLoad()} and the exported
\code{reconnect()} function (which closes existing connections before
calling \code{set_connections()}).
}
