% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{set_default_conn_name}
\alias{set_default_conn_name}
\title{Get/Set the default connection name}
\usage{
set_default_conn_name(new_default_name)
}
\arguments{
\item{new_default_name}{The name of a connection}
}
\description{
Get/Set the default connection name
}
