% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlrunners.R
\name{showsql}
\alias{showsql}
\title{Show to-be-executed SQL code}
\usage{
showsql(sql, is.queries = FALSE, interpolate = parent.frame())
}
\arguments{
\item{sql}{A character vector containing either sql file names or queries}

\item{is.queries}{Are these queries or filenames? defaults to FALSE (i.e. filenames)}

\item{interpolate}{An env containing parameters to interpolate into the sql, or NULL}
}
\value{
\code{showsql} returns one of:
\describe{
\item{\strong{A list of lists} (for multiple files)}{The outer list contains one list for each file in filenames. Each inner list contains interpolated queries from that file.}
\item{\strong{A list} (for single files containing several queries, or vectors of queries)}{Each element is an interpolated query.}
\item{\strong{A single element character vector} (for single queries or files containing a single query).}{A single interpolated SQL query}
}
}
\description{
Allows inspection of interpolated SQL before it is submitted to the database
}
\seealso{
Other SQL runners: 
\code{\link{runfiles}()},
\code{\link{runqueries}()}
}
\concept{SQL runners}
