% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{validate_config}
\alias{validate_config}
\title{Perform minimal sanity checks on a configuration}
\usage{
validate_config(conf)
}
\arguments{
\item{conf}{a single connection configuration returned by \link{get_all_configs}}
}
\value{
A valid connection configuration, or NULL
}
\description{
Perform minimal sanity checks on a configuration
}
\details{
\code{validate_conf} attempts to ensure that connection
configurations contain at least the elements \code{c(pool=boolean(),
  connection$Server=character())}, and that all child elements of
\code{connection} are character vectors or length 1.

If the \code{pool} element is not present, or not a logical of length 1, it
is inserted with the default value (FALSE).

If the \code{connection} element does not contain a \code{server} child
element, or a child element of \code{connection} is not a character vector
of length 1, NULL is returned.
}
