% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_configs.R
\name{validate_configs}
\alias{validate_configs}
\title{Apply validate_config to all connection configurations}
\usage{
validate_configs(configs)
}
\arguments{
\item{configs}{a configuration object returned by \link{read_all_configs}}
}
\value{
A list of valid configs, similar to one returned by \link{read_all_configs}
}
\description{
Apply validate_config to all connection configurations
}
\details{
Each configuration is passed to \link{validate_config}; invalid configs
are dropped with a warning
}
