% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlgenerators.R
\name{vinnthoffk}
\alias{vinnthoffk}
\title{Restrict a query to values found in a 1/nth sample of the unique values of a foreign key}
\usage{
vinnthoffk(v, n, f, k)
}
\arguments{
\item{v}{variable to restrict}

\item{n}{nth of fk}

\item{f}{foreign table}

\item{k}{foreign key}
}
\value{
string (sql code)
}
\description{
Restrict a query to values found in a 1/nth sample of the unique values of a foreign key
}
\seealso{
Other SQL Generators: 
\code{\link{nthofk}()}
}
\concept{SQL Generators}
