% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{yml2conn_str}
\alias{yml2conn_str}
\title{Convert a list object returned by read_yaml() to a db
connection string.}
\usage{
yml2conn_str(params)
}
\arguments{
\item{params}{Connection parameters}
}
\value{
Connection string

YAML values need to be double quoted in the YAML file.
}
\description{
Convert a list object returned by read_yaml() to a db
connection string.
}
