% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{yml2str}
\alias{yml2str}
\title{yml2str converts a list object returned by read_yaml() to a connection string.}
\usage{
yml2str(.conparms)
}
\arguments{
\item{.conparms}{Connection parameters}
}
\value{
Connection string

YAML values need to be double quoted in the YAML file.

For internal use.
}
\description{
yml2str converts a list object returned by read_yaml() to a connection string.
}
