% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_utils.R
\name{add_connection}
\alias{add_connection}
\title{Add a new connection to the connections cache}
\usage{
add_connection(conn_name, params)
}
\arguments{
\item{conn_name}{A string identifier for the new connection}

\item{params}{Connection parameters}
}
\description{
Add a new connection to the connections cache
}
