% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_configs.R
\name{combine_configs}
\alias{combine_configs}
\title{Combine optionally nested yaml config lists}
\usage{
combine_configs(root, new)
}
\arguments{
\item{root}{An yml-derived config list, as returned by
\code{\link[yaml:read_yaml]{yaml::read_yaml()}}}

\item{new}{Another yml-derived config list, to be inserted}
}
\value{
Combined yml config
}
\description{
Combine optionally nested yaml config lists
}
\details{
\itemize{
\item Elements that exists in new but not root will be added
\item Elements that exists in root but not new will be retained
\item Elements that exist in both will be replaced in root by the contents of new
}
}
