% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_info.R
\name{connection_info}
\alias{connection_info}
\title{Get information about available database connection}
\usage{
connection_info(name_str = ".*")
}
\arguments{
\item{name_str}{A regular expression to be used identify connection names to
include. The default ('.*') returns all of them.}
}
\value{
Null, or a tibble with 1 row per connection and the following fields:

\describe{
\item{name}{identifier (character)}
\item{description}{a description of the connection, if found in the conf file (character)}
\item{live}{is this connection valid and live? (logical)}
\item{driver}{the name of the driver function (character)}
\item{conn_str}{the string used to parameterize the connection (character)}
\item{pool}{is this a pool connection? (logical)}
}

If no connection names matched \code{name_str}, the tibble will be empty. If
no connections have been configured, \code{NULL} is returned.
}
\description{
Get information about available database connection
}
