% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect.R
\name{disconnect}
\alias{disconnect}
\title{Close all connections and remove them from the connections cache}
\usage{
disconnect()
}
\description{
Close all connections and remove them from the connections cache
}
