% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{driver}
\alias{driver}
\title{Determine the connection driver}
\usage{
driver(conf)
}
\arguments{
\item{conf}{A named list representing a single connection returned by
\code{\link{get_all_configs}}.}
}
\value{
A driver function. Defaults to \code{\link[odbc]{odbc}} if no
  "database_type" element is found.
}
\description{
Determine the connection driver
}
\details{
Search for an element named "database_type" and return a driver
  function appropriate for that database. For example, "database_type:
  sqlite" in your config will case \code{driver()} to return
  \code{\link[RSQLite]{SQLite}}.
}
