% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{live_connection}
\alias{live_connection}
\title{Return the named connection or NULL}
\usage{
live_connection(con_name)
}
\arguments{
\item{con_name}{Then name of the live connection you want (use
connections_list)}
}
\description{
Return the named connection or NULL
}
