% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{live_connection}
\alias{live_connection}
\title{Return the named connection or NULL}
\usage{
live_connection(con_name)
}
\arguments{
\item{con_name}{Then name of the live connection you want (use
connections_list to get names of available connetions)}
}
\value{
A live connection to a database, or NULL

  If \code{con_name} is not the name of a live connection to a database, a
  warning is issued and NULL is returned.
}
\description{
Return the named connection or NULL
}
