% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{read_config_file}
\alias{read_config_file}
\title{Return the contents of a config file}
\usage{
read_config_file(conf = NA)
}
\arguments{
\item{conf}{A string. either the full path and name of a config file, or one
of 'site' or 'user', or \code{NA} (the default). If \code{NA},
\code{get_config} will search the current directory for a config file.}
}
\value{
optionally nested named list or vector as returned by
  \code{yaml::read_yaml}
}
\description{
Return the contents of a config file
}
