% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_configs.R
\name{read_configs}
\alias{read_configs}
\title{Return the combined available configurations}
\usage{
read_configs(config_filename = NA, exclusive = FALSE)
}
\arguments{
\item{config_filename}{String. The full name and path of a configuration
file, or \code{NA} (the default).}

\item{exclusive}{Logical. If \code{TRUE}, the file named by \code{config_filename} is
treated as the only config file. Site and user level files are not read.
This parameter is ignored if \code{config_filename} is missing.}
}
\value{
optionally nested named list or vector as returned by
\code{\link[yaml:read_yaml]{yaml::read_yaml()}}
}
\description{
Reads and combines configuration files for database connections. By default, configuration files are
sought in user- and site-level config directories, and named by \code{config_filename}.
}
\details{
Reads and combines all available config files. User- and Site-level
files are sought in \code{\link[rappdirs:user_data_dir]{rappdirs::user_config_dir()}} and
\code{\link[rappdirs:site_data_dir]{rappdirs::site_config_dir()}} respectively, unless \code{exclusive=TRUE}.
}
