% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{read_sql}
\alias{read_sql}
\title{Read a sql file and return it's contents as a tibble}
\usage{
read_sql(file_name)
}
\arguments{
\item{file_name}{Full name and path of file to read}
}
\value{
A tibble containing 1 row per query and the following fields:
\describe{
\item{qname}{character. A name for this query}
\item{quotesql}{"yes" or "no". Should parameterized character values be quoted for this query? Defaults to "yes".}
\item{interpolate}{"yes" or "no". Should this query be parameterized with values from R? Defaults to "yes".}
\item{execmethod}{The method to execute this query.
One of "get" (\code{\link[DBI:dbGetQuery]{DBI::dbGetQuery()}}), "execute" (\code{\link[DBI:dbExecute]{DBI::dbExecute()}}), "sendq" (\code{\link[DBI:dbSendQuery]{DBI::dbSendQuery()}}), "sends" (\code{\link[DBI:dbSendStatement]{DBI::dbSendStatement()}}) or "spatial" (\code{\link[=sf:st_read]{sf:st_read()}})}
\item{geometry}{character. If \code{execmethod} is "spatial", which is the geometry column?}
\item{conn}{The name of the database connection to use for this query}
\item{sql}{The unparameterized sql query to be executed}
\item{filename}{The value of \code{file_name}}
}
}
\description{
Read a sql file and return it's contents as a tibble
}
