% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlrunners.R
\name{runfiles}
\alias{runfiles}
\title{Read, prepare and execute .SQL files}
\usage{
runfiles(
  filenames,
  quotesql = TRUE,
  values = parent.frame(),
  execmethod = "get",
  geometry = NA,
  default_conn = live_connection(get_default_conn_name()),
  include_params = FALSE
)
}
\description{
Accepts a character vector of SQL file names and attempts to execute each one.
}
