% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlrunners.R
\name{runqueries}
\alias{runqueries}
\title{Execute a sequence of SQL queries}
\usage{
runqueries(
  sql,
  quotesql = TRUE,
  values = parent.frame(),
  execmethod = "get",
  geometry = NA,
  default_conn = live_connection(get_default_conn_name()),
  include_params = FALSE
)
}
\arguments{
\item{sql}{An optionally-named list or character vector containing sql
commands, or a tibble returned by \code{\link{read_sql_file}}}

\item{quotesql}{"yes" or "no" - should interpolated characters be quoted by
default?}

\item{execmethod}{One of "get", "send" or "spatial" - which method should be
used to execute the query? "get" means \code{DBI::dbGetQuery}; "send" means
\code{DBI::dbSendQuery}; "spatial" means \code{sf::st_read}.}

\item{geometry}{If \code{execmethod} is "spatial", which geometry column
should be used (ignored if \code{execmethod} is not spatial)}

\item{include_params}{Should the parameters be included in the output?}

\item{interpolate}{Should the SQL be parameterized from R? Defaults to the
value of \code{parent.frame()}. Pass any object that is not an environment
(e.g. "no" or FALSE) if interpolation is to be skipped, or another
environment containing values to interpolate to avoid using
\code{.GlobalEnv}.}

\item{conn}{Either the name of a sqlhelper connection, or a database
connection returned by \code{DBI::dbConnect}, or NA}
}
\value{
If \code{include_params} is \code{FALSE}, an named list containing
the results of each query; the names are the same as those in the \code{sql}
parameter. If \code{include_params} is \code{TRUE}, a tibble containing the query as executed, the above parameters,
and the result of each query.
}
\description{
Accepts a character vector of SQL queries and runs each one
}
\seealso{
\code{\link{runfiles}}

Other SQL runners: 
\code{\link{runfiles}()}
}
\concept{SQL runners}
