% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlrunners.R
\name{runqueries}
\alias{runqueries}
\title{Execute a sequence of SQL queries}
\usage{
runqueries(db, queries, interpolate = parent.frame())
}
\arguments{
\item{db}{Which database do you want to query? One of <'hive' or 'h'> for Hive
or <'pg', 'postgres', postgresql', or 'p'> for PostgreSQL.}

\item{queries}{A list or character vector containing your sql commands}

\item{interpolate}{Should the SQL be parameterized from R? Defaults to the
value of \code{parent.frame()}. May be set to \code{FALSE} if interpolation is
to be avoided, or to another environment to control the source of the used.
This can be useful if UC SQL Helper is used to run SQL from within another
package, and you don't want to interfere with \code{.GlobalEnv}. The
\code{\link[uccaseload]{UC Caseload package}} contains examples of this.}
}
\value{
A list containing the results of each query
}
\description{
Accepts a character vector of SQL queries and runs each one
on either hive or postgresql.
}
\seealso{
\code{\link{runfiles}} \code{\link{rundirs}}

Other SQL runners: 
\code{\link{runfiles}()},
\code{\link{showsql}()}
}
\concept{SQL runners}
