% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_default_conn_name.R
\name{set_default_conn_name}
\alias{set_default_conn_name}
\alias{get_default_conn_name}
\title{Set the name of the default connection to use}
\usage{
set_default_conn_name(conn_name)

get_default_conn_name()
}
\arguments{
\item{conn_name}{Character string. The name a connection}
}
\value{
The name of the default connection
}
\description{
Set the name of the default connection to use

Get the name of the default connection
}
