% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlhelper.R
\docType{package}
\name{sqlhelper}
\alias{sqlhelper}
\title{SQL Helper}
\description{
Easier SQL interaction with data.
}
\details{
The \code{sqlhelper} package does three things:

\describe{

\item{\strong{It manages connections to
databases}.}{They are opened automatically when the package is loaded, and
closed when it is unloaded.}

\item{\strong{It provides functions for running sql queries and files of sql
queries}.}{You simply provide a query or filename (or list of queries or
filenames). See \code{\link{runfiles}}.}

\item{\strong{It provides functions which return useful sql snippets}.}{For example,
for use in a SQL \code{WHERE} clause to extract an \code{n} percent sample
from a table. See, \code{\link{nthofk}}.}

}
}
\section{}{
NA
}

\author{
Matthew Roberts, \email{matthew.roberts@communities.gov.uk}
}
