% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{yml2conn_str}
\alias{yml2conn_str}
\title{yml2conn_str converts a list object returned by read_yaml() to a db
connection string.}
\usage{
yml2conn_str(parms)
}
\arguments{
\item{parms}{Connection parameters}
}
\value{
Connection string

YAML values need to be double quoted in the YAML file.
}
\description{
yml2conn_str converts a list object returned by read_yaml() to a db
connection string.
}
